POL-166: move formatDate to dateutils
