+++ src/pages/timelines.js
@@ -5,7 +5,7 @@ import Layout from "../components/layout"
 import {getTimelinesDescription, getTimelinesTitle} from "../utils/seoutils";
 import {Container} from "../components/container";
 import {TimelineCalendar} from "../components/calendar";
-import {formatDate} from "../utils/formatutils"
+import {formatDate} from "../utils/dateutils"
 import {CALENDAR_TIMESTAMP_KEY} from "../utils/constants";
