+++ src/templates/bill.js
@@ -7,7 +7,7 @@ import LinkCard from "../components/linkCard"
 import Layout from "../components/layout"
 import SEO from "../components/seo"
 import MinutesCard from "../components/minutesCard"
-import {formatDate} from "../utils/formatutils"
+import {formatDate} from "../utils/dateutils"
 import {buildPath} from "../utils/urlutils";
 import {getBillDescription} from "../utils/seoutils";
 import {sortMinutesList, sortNewsList} from "../utils/sortutils";
